Implement a first versions of Container.run()
It's pretty limited right now because you can't stream the output of the
command run in the container in real time. The reason behind that
limitation is that you can't safely get the id of a docker container and
its stdout/stderr at the same time.

This limitation probably won't be lifted until Docker gets a real API
(this is code using the CLI for now).

Container.run() is implemented as a context manager which returns a
subprocess.Popen-like object. (it's actually a subprocess.Popen object
but this will change if we stop using the CLI). It allows you to write
things like this:

i = Image(ImageRevSpec.parse("lopter/raring-base:latest"))
c = i.instantiate()
with c.run(["cat"], stdin=c.PIPE) as cat:
    cat.stdin.write("Mouse\n")
    cat.stdin.write("Meow?\n")
    cat.stdin.close()
    # The idea is also to be able to use cat.stdout
    # and cat.stderr but it's not possible right now
print c.logs
i = c.result