Unpack the service and application tarballs in the container
This also add the file `definition.json` in the service tarball, as it
contains a lot of useful informations to build the service (config,
extra requirements, approot…) that we don't want to parse again. It will
help to move the supervisor/environment files generation from the
sandbox package to the builder package too (where I think it makes more
sense).