Speedup the builder bootstrap (was a regression from #2)
And unwanted side-effect of 9002f19d25bcc785b4e2a4275feee58ae4e91d76 was
to install the sandbox's dependencies inside the build container, where
they aren't needed since we only use the "builder" part.

This changeset modifies bootstrap.sh to use pip install --no-deps and to
install builder's dependencies manually. It also changes dotcloud-builder
to be a script in setup.py instead of an entry_point (because setuptools
checks that all the requirements are here from the "entry-point" script
it generates…). It sucks to not be able to manage that from setup.py
(using extra_require [1] doesn't seem to be a viable option), but it
sucks less than a slow build or a two packages instead of one (see #2).

[1] http://peak.telecommunity.com/DevCenter/setuptools#declaring-extras-optional-features-with-their-own-dependencies